+++ docs/CHANGELOG.md
@@ -25,8 +25,8 @@ The format is based on [Keep a Changelog][kac] and this project adheres to
   folder name to allow for multilib install, e.g. into lib64 (#452)
 * add `run` flag `--keep-empty-lines` to retain empty lines in `${lines[@]}` (#224,
   a894fbfa)
-* add `run` flag `--output merged|separate|stderr|stdout` (default: `merged`) to
-  select how `$output` and `$lines` are filled (#47, 5c9b173d)
+* add `run` flag `--separate-stderr` which also fills `$stderr` and
+  `$stderr_lines` (#47, 5c9b173d, #507)
 
 ### Fixed